extract GcsStorage interface, provide a noop data plane implementation
