+++ service/src/main/java/org/databiosphere/workspacedataservice/storage/GcsStorageConfig.java
@@ -3,16 +3,24 @@ package org.databiosphere.workspacedataservice.storage;
 import com.google.cloud.spring.core.GcpProjectIdProvider;
 import java.io.IOException;
 import org.databiosphere.workspacedataservice.annotations.DeploymentMode.ControlPlane;
+import org.databiosphere.workspacedataservice.annotations.DeploymentMode.DataPlane;
 import org.databiosphere.workspacedataservice.config.DataImportProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ControlPlane
 public class GcsStorageConfig {
+
+  @ControlPlane
   @Bean
   public GcsStorage gcsStorage(
       GcpProjectIdProvider projectIdProvider, DataImportProperties properties) throws IOException {
-    return GcsStorage.create(projectIdProvider.getProjectId(), properties);
+    return GcsStorageImpl.create(projectIdProvider.getProjectId(), properties);
+  }
+
+  @DataPlane
+  @Bean
+  public GcsStorage noopGcsStorage() throws IOException {
+    return new GcsStorageNoopImpl();
   }
 }
